Feature - Save discount values on sales table

+++ controllers/DiscountController.js
@@ -1,6 +1,7 @@
 const SaleRepository = require("../repositories/SaleRepository");
 const DiscountRepository = require("../repositories/DiscountRepository");
 const { handleError, successResponse } = require("../utils/utils");
+const { formatToDecimal } = require("../helpers/util");
 
 const getAllDiscountsBySale = async (req, res) => {
     try {
@@ -12,14 +13,20 @@ const getAllDiscountsBySale = async (req, res) => {
 }
 
 const createDiscount = async (req, res) => {
-    try {
-        const sale = await SaleRepository.findByPk(req.body.sale_id);
+    try 
+    {
+        const { sale_id, discount, discount_converted } = req.body;
+        const sale = await SaleRepository.findByPk(sale_id);
 
         req.body.total_amount_sale = sale.total_amount;
         req.body.total_amount_sale_converted = sale.total_amount_converted;
 
-        const discount = await DiscountRepository.create(req.body);
-        successResponse( res, { discount: discount } );
+        const created = await DiscountRepository.create(req.body);
+
+        await SaleRepository.updateDiscountAmount(sale_id, discount, discount_converted);
+
+        successResponse( res, { discount: created } );
+
     } catch (error) {
         handleError( res, error );
     }
@@ -27,8 +34,25 @@ const createDiscount = async (req, res) => {
 
 const deleteDiscount = async (req, res) => {
     try {
-        const destroyed = await DiscountRepository.destroy(req.params.id);
-        successResponse( res, { discount: destroyed } );
+        
+        const discount = await DiscountRepository.findById(req.params.id);
+
+        if(discount)
+        {
+            const sale = await SaleRepository.findByPk(discount.sale_id);
+
+            sale.discount_amount = Number(sale.discount_amount)-Number(discount.discount);
+            sale.discount_amount_converted = Number(sale.discount_amount_converted)-Number(discount.discount_converted);
+            sale.total_amount = Number(sale.total_amount)+Number(discount.discount);
+            sale.total_amount_converted = Number(sale.total_amount_converted)+Number(discount.discount_converted);
+            
+            sale.save();
+
+            const destroyed = await DiscountRepository.destroy(req.params.id);
+            successResponse( res, { discount: destroyed } );
+        }
+
+
     } catch (error) {
         handleError( res, error );
     }
